requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.5.0=h63175ca_0
blas=1.0=mkl
blas-devel=3.9.0=12_win64_mkl
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
brotlipy=0.7.0=py310h8d17308_1005
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.12.7=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.16.0=hd694305_1014
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h628cb3f_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=py310h232114e_0
cryptography=39.0.1=py310h6e82f81_0
cudatoolkit=11.8.0=h09e9e62_11
cycler=0.11.0=pyhd8ed1ab_0
expat=2.5.0=h1537add_0
ffmpeg=5.1.2=gpl_h5b1d025_106
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=py310h8d17308_1
freeglut=3.2.2=h0e60522_1
freetype=2.12.1=h546665d_1
future=0.18.3=pyhd8ed1ab_0
gettext=0.21.1=h5728263_0
glib=2.74.1=h12be248_1
glib-tools=2.74.1=h12be248_1
graphite2=1.3.13=1000
gst-plugins-base=1.22.0=h001b923_0
gstreamer=1.22.0=h6b5321d_0
h5py=3.8.0=nompi_py310h727aef9_101
harfbuzz=6.0.0=he256f1b_0
hdf5=1.14.0=nompi_h5af7005_102
icu=70.1=h0e60522_0
idna=3.4=pyhd8ed1ab_0
intel-openmp=2023.0.0=h57928b3_25922
jasper=2.0.33=hc2e4405_1
jpeg=9e=h8ffe710_2
kiwisolver=1.4.4=py310h232114e_1
krb5=1.20.1=heb0366b_0
lcms2=2.14=h90d422f_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=12_win64_mkl
libclang=15.0.7=default_h77d9078_1
libclang13=15.0.7=default_h77d9078_1
libcurl=7.87.0=h68f0423_0
libdeflate=1.14=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libglib=2.74.1=he8f3873_1
libhwloc=2.8.0=h039e092_1
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=12_win64_mkl
liblapacke=3.9.0=12_win64_mkl
libogg=1.3.4=h8ffe710_1
libopencv=4.5.3=py310h8245ef6_8
libopus=1.3.1=h8ffe710_1
libpng=1.6.39=h19919ed_0
libprotobuf=3.21.12=h12be248_0
libsqlite=3.40.0=hcfcfb64_0
libssh2=1.10.0=h9a1e1f7_3
libtiff=4.4.0=hc4f729c_5
libuv=1.44.2=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libxml2=2.10.3=hc3477c8_0
libzlib=1.2.13=hcfcfb64_4
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.6.3=py310h5588dad_0
matplotlib-base=3.6.3=py310h51140c5_0
mkl=2021.4.0=h0e2418a_729
mkl-devel=2021.4.0=h57928b3_730
mkl-include=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py310hcf6e17e_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
ninja=1.11.0=h2d74725_0
numpy=1.24.2=py310hd02465a_0
opencv=4.5.3=py310h5588dad_8
openh264=2.3.1=h63175ca_2
openjpeg=2.5.0=hc9384bd_1
openssl=3.0.8=hcfcfb64_0
packaging=23.0=pyhd8ed1ab_0
pcre2=10.40=h17e33f8_0
pillow=9.2.0=py310hd4fb230_3
pip=23.0=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
ply=3.11=py_1
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
py-opencv=4.5.3=py310hbbfc1a7_8
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py310h1fd54f2_3
pyqt5-sip=12.11.0=py310h00ffb61_3
pysocks=1.7.1=pyh0701188_6
python=3.10.9=h4de0772_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=1.12.1=cpu_py310h5e1f01c_0
pytorch-mutex=1.0=cpu
pyyaml=6.0=py310h8d17308_5
qt-main=5.15.8=h720456b_6
requests=2.28.2=pyhd8ed1ab_0
setuptools=67.1.0=pyhd8ed1ab_0
sip=6.7.7=py310h00ffb61_0
six=1.16.0=pyh6c4a22f_0
svt-av1=1.4.1=h63175ca_0
tbb=2021.7.0=h91493d7_1
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
torchvision=0.13.1=cpu_py310h378ed51_0
tornado=6.2=py310h8d17308_1
tqdm=4.64.1=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.0.0=py310h8d17308_0
urllib3=1.26.14=pyhd8ed1ab_0
vc=14.3=hb6edc58_10
vs2015_runtime=14.34.31931=h4c5c07a_10
wheel=0.38.4=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h12be248_6
